requirements.txt
asgiref==3.8.1
Django==5.1
sqlparse==0.5.1

resend~=0.8.0
python-dotenv~=1.0.1